+++ src/components/AgreementForm/Landlord.tsx
@@ -1,5 +1,4 @@
 import React from "react";
-import { Field } from "formik";
 import { Button } from "@chakra-ui/core";
 
 import { useHistory } from "react-router-dom";
