+++ src/components/AgreementForm/Signatures.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 import moment from "moment";
-import { FieldArray, useFormikContext } from "formik";
+import { FieldArray } from "formik";
 
-import { FormValues } from "../../interfaces";
+// import { FormValues } from "../../interfaces";
 import { FormikSingleDatePicker } from "../FormikDates";
 import { displayFullName } from "../../helpers/functions";
 import FieldSet from "../FieldSet";
