+++ src/pages/Agreement.tsx
@@ -1,10 +1,9 @@
 import React from "react";
-import moment from "moment";
 import { Switch, Route, Link, Redirect } from "react-router-dom";
 import { FormikProps, withFormik } from "formik";
-import { EditorState } from "draft-js";
 
 // import { FormValues } from "../interfaces";
+import initialValues from "../components/AgreementForm/initialValues";
 import validationSchema from "../components/AgreementForm/validationSchema";
 
 import Title from "../components/AgreementForm/Title";
@@ -16,87 +15,12 @@ import Rent from "../components/AgreementForm/RentAndDeposit/Rent";
 import SecurityDeposit from "../components/AgreementForm/RentAndDeposit/SecurityDeposit";
 import BillsUtilities from "../components/AgreementForm/BillsUtilities";
 import Signatures from "../components/AgreementForm/Signatures";
-import TestDraft from "../components/AgreementForm/TestDraft";
+// import TestDraft from "../components/AgreementForm/TestDraft";
 
 import "draft-js/dist/Draft.css";
 
-// function to flatten array for field array: https://github.com/jaredpalmer/formik/issues/11
-
-const billShape = {
-  name: null,
-  totalAmt: 0,
-  dueDate: moment(),
-  interval: [] // once, monthly, every 2 months, annually
-};
-
 const formikEnhancer = withFormik({
-  mapPropsToValues: props => ({
-    landlord: {
-      firstName: "",
-      lastName: "",
-      address: "",
-      phone: "",
-      email: "",
-      company: ""
-    },
-    household: {
-      address: "",
-      leaseDates: {
-        startDate: moment(),
-        endDate: moment()
-      },
-      smokingAllowed: false,
-      petFriendly: true,
-      bedroomsAmt: 2,
-      bathroomsAmt: 1,
-      rentAmt: "",
-      securityDepositAmt: ""
-    },
-    roommates: [
-      { firstName: "Roommate", lastName: "One", email: "roomie1@email.com", phone: "6041234567" },
-      { firstName: "", lastName: "", email: "", phone: "" }
-      // {
-      //   firstName: "Roommate",
-      //   lastName: "Twwwooooo",
-      //   email: "bestestroommate@email.com",
-      //   phone: "7781234567"
-      // }
-    ],
-    rent: {
-      ...billShape,
-      portion: [{ roommate: [], roommate_amt: 0, amt_type: [] }]
-    },
-    securityDeposit: {
-      ...billShape,
-      portion: [{ roommate: [], roommate_amt: 0, amt_type: [] }]
-    },
-    bills: [{ ...billShape }],
-    housekeeping: {
-      weekdayPM: "",
-      weekdayAM: "",
-      weekendPM: "",
-      weekendAM: "",
-      guestPolicy: EditorState.createEmpty(),
-      spacesPolicy: EditorState.createEmpty(),
-      roomsPolicy: EditorState.createEmpty(),
-      choresPolicy: EditorState.createEmpty(),
-      vacationPolicy: EditorState.createEmpty(),
-      personalItemsPolicy: EditorState.createEmpty(),
-      smokingPolicy: EditorState.createEmpty(),
-      messagesPolicy: EditorState.createEmpty(),
-      petsPolicy: EditorState.createEmpty()
-    },
-    signatures: [{ fullName: "", agreed: false, date: moment() }],
-    // test values for TestDraft.tsx
-    textArea1: EditorState.createEmpty(),
-    textArea2: EditorState.createEmpty(),
-    status: [],
-    leaseDates: {
-      startDate: null,
-      endDate: null
-    },
-    billDate: moment()
-  }),
+  mapPropsToValues: props => initialValues,
   handleSubmit: (values, { setSubmitting }) => {
     alert("it submitted! check console for values!");
     console.log("form submission values", values);
@@ -216,9 +140,9 @@ const Agreement = () => {
         <li>
           <Link to="/agreement/signatures">Signatures</Link>
         </li>
-        <li>
+        {/* <li>
           <Link to="/agreement/testDraft">Third Party Examples</Link>
-        </li>
+        </li> */}
       </ul>
       {/* enhanced agreement form */}
       <EnhancedAgreement />
